refactor: loginの処理をvuexに移動する
