+++ client/app/pages/signin.vue
@@ -6,9 +6,7 @@
 import Vue from 'vue'
 import { mapActions } from 'vuex'
 import GranLogin from '~/components/templates/GranLogin.vue'
-import firebase from '~/plugins/firebase'
 import { ILoginForm } from '~/types/form'
-import { IUser } from '~/types/store'
 
 export default Vue.extend({
   components: {
@@ -19,24 +17,13 @@ export default Vue.extend({
     isError: false
   }),
   methods: {
-    ...mapActions({
-      setUser: 'auth/setUser'
-    }),
+    ...mapActions('auth', ['loginWithEmailAndPassword']),
     async login(loginForm: ILoginForm) {
-      await firebase
-        .auth()
-        .signInWithEmailAndPassword(loginForm.email, loginForm.password)
-        .then((auth: any) => {
-          const user: IUser = {
-            uid: auth.user.uid,
-            email: auth.user.email,
-            creationTime: auth.user.metadata.creationTime,
-            lastSignInTime: auth.user.metadata.lastSignInTime
-          }
-          this.setUser(user)
+      await this.loginWithEmailAndPassword(loginForm)
+        .then(() => {
+          this.$router.push('/inspire') // 仮のルーティング
         })
         .catch(() => {
-          // firebaseとの接続エラーだとまずいな
           this.isError = true
         })
     },
